Combined the two main.js
moved everything form components/main.js to ./main.js
